backend/requirements.txt
asgiref==3.8.1
Django==3.2.23
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-nested-routers==0.93.5
pytz==2024.1
sqlparse==0.4.4
